added files for bf16 support

+++ trojai-llm-mitigation-framework/trojai_llm_mitigation_round/mitigations/finetuning.py
@@ -12,8 +12,8 @@ from ..utils import print_summary
 
 
 class FineTuningTrojaiMitigationLLM(TrojAIMitigationLLM):
-    def __init__(self, lr, train_epochs, optim, device, batch_size=4, num_workers=1, fp16=False, **kwargs):
-        super().__init__(device, batch_size, num_workers, fp16, **kwargs)
+    def __init__(self, lr, train_epochs, optim, device, batch_size=4, num_workers=1, bf16=False, **kwargs):
+        super().__init__(device, batch_size, num_workers, bf16, **kwargs)
         self.lr = lr
         self.train_epochs = train_epochs
         self.optim = optim
@@ -26,7 +26,6 @@ class FineTuningTrojaiMitigationLLM(TrojAIMitigationLLM):
             target_token_length = self.max_token_length        
 
         peft_model = get_peft_model(model, peft_config)
-
         training_args = TrainingArguments(output_dir="test_trainer",
                                             learning_rate=self.lr,
                                             num_train_epochs=self.train_epochs,
@@ -34,7 +33,7 @@ class FineTuningTrojaiMitigationLLM(TrojAIMitigationLLM):
                                             evaluation_strategy="epoch", 
                                             per_device_train_batch_size=self.batch_size,
                                             per_device_eval_batch_size=self.batch_size,
-                                            fp16=self.fp16,
+                                            bf16=self.bf16,
                                             logging_strategy="epoch",
                                             report_to="tensorboard")
         trainer = SFTTrainer(
@@ -45,6 +44,7 @@ class FineTuningTrojaiMitigationLLM(TrojAIMitigationLLM):
             data_collator=collator,
             max_seq_length=self.max_token_length
         )
+        
         print(f"Beginning Training with {len(dataset['train'])} examples")
         result = trainer.train()
         print_summary(result)
